Move full configuration logic from `rustc_expand` to `rustc_builtin_macros`
This logic is applicable to two specific macros and not to the expansion infrastructure in general.
